Refactor fieldset errors to use aria-describedby on the fieldset rather than individual fields AB#143786 AB#137104
