+++ GovUk.Frontend.Umbraco.ExampleApp/Models/ErrorMessageViewModel.cs
@@ -13,8 +13,6 @@ namespace GovUk.Frontend.Umbraco.ExampleApp.Models
         [Required(ErrorMessage = nameof(Field2))]
         public int? Field2 { get; set; }
 
-        public int? Field3 { get; set; }
-
         [Range(100, 100, ErrorMessage = nameof(AddsUpTo100))]
         public int? AddsUpTo100 { get => Field1 + Field2; }
