+++ GovUk.Frontend.Umbraco.ExampleApp/Models/ModelsBuilder/GovukFieldsetSettings.generated.cs
@@ -50,11 +50,11 @@ namespace Umbraco.Cms.Web.Common.PublishedModels
 		// properties
 
 		///<summary>
-		/// Render error classes: When a nested 'Error message' component is bound to a property, render GOV.UK error classes when that property fails validation.
+		/// Fieldset errors enabled: When a child 'Error message' component is bound to a property on the view model which fails validation, that is a fieldset error.
 		///</summary>
 		[global::System.CodeDom.Compiler.GeneratedCodeAttribute("Umbraco.ModelsBuilder.Embedded", "10.0.0+e3f4b86")]
-		[ImplementPropertyType("renderErrorClasses")]
-		public virtual bool RenderErrorClasses => this.Value<bool>(_publishedValueFallback, "renderErrorClasses");
+		[ImplementPropertyType("fieldsetErrors")]
+		public virtual bool FieldsetErrors => this.Value<bool>(_publishedValueFallback, "fieldsetErrors");
 
 		///<summary>
 		/// CSS classes
