+++ GovUk.Frontend.Umbraco.ExampleApp/uSync/v9/Content/error-message.config
@@ -66,7 +66,7 @@
     {
       "contentTypeKey": "f7fc1212-5317-4642-9565-f8b6585508e8",
       "udi": "umb://element/c490526fcfe4415eac16884b2870132c",
-      "blocks": "{\r\n  \"layout\": {\r\n    \"Umbraco.BlockList\": [\r\n      {\r\n        \"contentUdi\": \"umb://element/0ddcf3cf0b8b45d7bf0df892e3b38913\",\r\n        \"settingsUdi\": \"umb://element/1f1395f71c0b4651962af6f9c47fabad\"\r\n      },\r\n      {\r\n        \"contentUdi\": \"umb://element/edb075ccdd9e4a159e894730ddae63cb\",\r\n        \"settingsUdi\": \"umb://element/2d3cd57250574fc7931bdeaf43a571eb\"\r\n      },\r\n      {\r\n        \"contentUdi\": \"umb://element/29fef1a65dd5463c8f72b5d4c15c65f8\",\r\n        \"settingsUdi\": \"umb://element/3e9257cf131d4cdd95e5f05f5c8c7877\"\r\n      },\r\n      {\r\n        \"contentUdi\": \"umb://element/7002b58da6134d27850bddf2c4d00b36\",\r\n        \"settingsUdi\": \"umb://element/f4f5ec3369ff4adba2f650b786f77080\"\r\n      }\r\n    ]\r\n  },\r\n  \"contentData\": [\r\n    {\r\n      \"contentTypeKey\": \"2cb36d91-7480-4234-a5d8-6961a03ada8a\",\r\n      \"udi\": \"umb://element/edb075ccdd9e4a159e894730ddae63cb\",\r\n      \"error\": \"Those two numbers do not add up to 100\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"fe2dd0de-55a9-4249-be03-6896aa719974\",\r\n      \"udi\": \"umb://element/29fef1a65dd5463c8f72b5d4c15c65f8\",\r\n      \"label\": \"Enter a number\",\r\n      \"hint\": \"\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"fe2dd0de-55a9-4249-be03-6896aa719974\",\r\n      \"udi\": \"umb://element/7002b58da6134d27850bddf2c4d00b36\",\r\n      \"label\": \"Enter another number\",\r\n      \"hint\": \"\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"2cb36d91-7480-4234-a5d8-6961a03ada8a\",\r\n      \"udi\": \"umb://element/0ddcf3cf0b8b45d7bf0df892e3b38913\",\r\n      \"error\": \"Those two numbers do not add up to 50\"\r\n    }\r\n  ],\r\n  \"settingsData\": [\r\n    {\r\n      \"contentTypeKey\": \"b34b41a7-261a-41a3-8877-afc4086cc065\",\r\n      \"udi\": \"umb://element/2d3cd57250574fc7931bdeaf43a571eb\",\r\n      \"cssClasses\": null,\r\n      \"modelProperty\": \"AddsUpTo100\",\r\n      \"columnSize\": null,\r\n      \"columnSizeFromDesktop\": null\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"98070ce9-d528-4bfc-bd90-e411e5110f56\",\r\n      \"udi\": \"umb://element/3e9257cf131d4cdd95e5f05f5c8c7877\",\r\n      \"modelProperty\": \"Field1\",\r\n      \"labelIsPageHeading\": 0,\r\n      \"columnSize\": null,\r\n      \"columnSizeFromDesktop\": null,\r\n      \"cssClasses\": \"govuk-input--width-3\",\r\n      \"errorMessageRequired\": \"Enter a number\",\r\n      \"validationProperty1\": \"AddsUpTo50\",\r\n      \"validationProperty2\": \"AddsUpTo100\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"98070ce9-d528-4bfc-bd90-e411e5110f56\",\r\n      \"udi\": \"umb://element/f4f5ec3369ff4adba2f650b786f77080\",\r\n      \"modelProperty\": \"Field2\",\r\n      \"labelIsPageHeading\": 0,\r\n      \"columnSize\": null,\r\n      \"columnSizeFromDesktop\": null,\r\n      \"cssClasses\": \"govuk-input--width-3\",\r\n      \"errorMessageRequired\": \"Enter a number\",\r\n      \"validationProperty1\": \"AddsUpTo50\",\r\n      \"validationProperty2\": \"AddsUpTo100\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"b34b41a7-261a-41a3-8877-afc4086cc065\",\r\n      \"udi\": \"umb://element/1f1395f71c0b4651962af6f9c47fabad\",\r\n      \"cssClasses\": null,\r\n      \"modelProperty\": \"AddsUpTo50\",\r\n      \"columnSize\": null,\r\n      \"columnSizeFromDesktop\": null\r\n    }\r\n  ]\r\n}",
+      "blocks": "{\r\n  \"layout\": {\r\n    \"Umbraco.BlockList\": [\r\n      {\r\n        \"contentUdi\": \"umb://element/0ddcf3cf0b8b45d7bf0df892e3b38913\",\r\n        \"settingsUdi\": \"umb://element/1f1395f71c0b4651962af6f9c47fabad\"\r\n      },\r\n      {\r\n        \"contentUdi\": \"umb://element/edb075ccdd9e4a159e894730ddae63cb\",\r\n        \"settingsUdi\": \"umb://element/2d3cd57250574fc7931bdeaf43a571eb\"\r\n      },\r\n      {\r\n        \"contentUdi\": \"umb://element/29fef1a65dd5463c8f72b5d4c15c65f8\",\r\n        \"settingsUdi\": \"umb://element/3e9257cf131d4cdd95e5f05f5c8c7877\"\r\n      },\r\n      {\r\n        \"contentUdi\": \"umb://element/7002b58da6134d27850bddf2c4d00b36\",\r\n        \"settingsUdi\": \"umb://element/f4f5ec3369ff4adba2f650b786f77080\"\r\n      }\r\n    ]\r\n  },\r\n  \"contentData\": [\r\n    {\r\n      \"contentTypeKey\": \"2cb36d91-7480-4234-a5d8-6961a03ada8a\",\r\n      \"udi\": \"umb://element/edb075ccdd9e4a159e894730ddae63cb\",\r\n      \"error\": \"Those two numbers do not add up to 100\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"fe2dd0de-55a9-4249-be03-6896aa719974\",\r\n      \"udi\": \"umb://element/29fef1a65dd5463c8f72b5d4c15c65f8\",\r\n      \"label\": \"Enter a number\",\r\n      \"hint\": \"\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"fe2dd0de-55a9-4249-be03-6896aa719974\",\r\n      \"udi\": \"umb://element/7002b58da6134d27850bddf2c4d00b36\",\r\n      \"label\": \"Enter another number\",\r\n      \"hint\": \"\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"2cb36d91-7480-4234-a5d8-6961a03ada8a\",\r\n      \"udi\": \"umb://element/0ddcf3cf0b8b45d7bf0df892e3b38913\",\r\n      \"error\": \"Those two numbers do not add up to 50\"\r\n    }\r\n  ],\r\n  \"settingsData\": [\r\n    {\r\n      \"contentTypeKey\": \"b34b41a7-261a-41a3-8877-afc4086cc065\",\r\n      \"udi\": \"umb://element/2d3cd57250574fc7931bdeaf43a571eb\",\r\n      \"cssClasses\": null,\r\n      \"modelProperty\": \"AddsUpTo100\",\r\n      \"columnSize\": null,\r\n      \"columnSizeFromDesktop\": null\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"98070ce9-d528-4bfc-bd90-e411e5110f56\",\r\n      \"udi\": \"umb://element/3e9257cf131d4cdd95e5f05f5c8c7877\",\r\n      \"modelProperty\": \"Field1\",\r\n      \"labelIsPageHeading\": 0,\r\n      \"columnSize\": null,\r\n      \"columnSizeFromDesktop\": null,\r\n      \"cssClasses\": \"govuk-input--width-3\",\r\n      \"errorMessageRequired\": \"Enter a number\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"98070ce9-d528-4bfc-bd90-e411e5110f56\",\r\n      \"udi\": \"umb://element/f4f5ec3369ff4adba2f650b786f77080\",\r\n      \"modelProperty\": \"Field2\",\r\n      \"labelIsPageHeading\": 0,\r\n      \"columnSize\": null,\r\n      \"columnSizeFromDesktop\": null,\r\n      \"cssClasses\": \"govuk-input--width-3\",\r\n      \"errorMessageRequired\": \"Enter a number\"\r\n    },\r\n    {\r\n      \"contentTypeKey\": \"b34b41a7-261a-41a3-8877-afc4086cc065\",\r\n      \"udi\": \"umb://element/1f1395f71c0b4651962af6f9c47fabad\",\r\n      \"cssClasses\": null,\r\n      \"modelProperty\": \"AddsUpTo50\",\r\n      \"columnSize\": null,\r\n      \"columnSizeFromDesktop\": null\r\n    }\r\n  ]\r\n}",
       "legend": "Enter two numbers that add up to 50 or 100"
     },
     {
@@ -94,11 +94,11 @@
     {
       "contentTypeKey": "6930e373-0fce-44ad-89b8-267442675807",
       "udi": "umb://element/0b7f8bfff4c543ebbe10b37df2627fdc",
-      "legendIsPageHeading": 0,
+      "legendIsPageHeading": 1,
       "cssClassesForRow": null,
       "columnSize": "[\"full\"]",
       "columnSizeFromDesktop": null,
-      "renderErrorClasses": 1
+      "fieldsetErrors": 1
     },
     {
       "contentTypeKey": "ec4a6e3f-a6a8-4c8a-9d4c-d3e90399ac2c",
