+++ GovUk.Frontend.Umbraco.ExampleApp/uSync/v9/ContentTypes/govukfieldsetsettings.config
@@ -1,55 +1,55 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <ContentType Key="6930e373-0fce-44ad-89b8-267442675807" Alias="govukFieldsetSettings" Level="2">
-  <Info>
-    <Name>Fieldset settings</Name>
-    <Icon>icon-equalizer color-black</Icon>
-    <Thumbnail>folder.png</Thumbnail>
-    <Description>Settings for the fieldset component. Use the fieldset component when you need to show a relationship between multiple form inputs.</Description>
-    <AllowAtRoot>False</AllowAtRoot>
-    <IsListView>False</IsListView>
-    <Variations>Nothing</Variations>
-    <IsElement>true</IsElement>
-    <HistoryCleanup>
-      <PreventCleanup>False</PreventCleanup>
-      <KeepAllVersionsNewerThanDays></KeepAllVersionsNewerThanDays>
-      <KeepLatestVersionPerDayForDays></KeepLatestVersionPerDayForDays>
-    </HistoryCleanup>
-    <Folder>GOV.UK</Folder>
-    <Compositions>
-      <Composition Key="bf69e8d6-97a6-4b50-bec7-c5993e120339">govukCssClasses</Composition>
-      <Composition Key="d508995b-a958-4273-8ab3-50735220a1b2">govukGrid</Composition>
-      <Composition Key="d7016563-0b6f-4c3a-b75d-b0bbe66419fd">govukGridColumnClasses</Composition>
-      <Composition Key="8ba759bf-131b-488e-b7e6-98ff91bb076b">govukLegendIsPageHeading</Composition>
-    </Compositions>
-    <DefaultTemplate></DefaultTemplate>
-    <AllowedTemplates />
-  </Info>
-  <Structure />
-  <GenericProperties>
-    <GenericProperty>
-      <Key>052c190a-53f3-400c-bbd7-41d69328aa35</Key>
-      <Name>Render error classes</Name>
-      <Alias>renderErrorClasses</Alias>
-      <Definition>189de753-cac2-4307-868b-114832c68a07</Definition>
-      <Type>Umbraco.TrueFalse</Type>
-      <Mandatory>false</Mandatory>
-      <Validation></Validation>
-      <Description><![CDATA[When a nested 'Error message' component is bound to a property, render GOV.UK error classes when that property fails validation.]]></Description>
-      <SortOrder>10</SortOrder>
-      <Tab Alias="settings">Settings</Tab>
-      <Variations>Nothing</Variations>
-      <MandatoryMessage></MandatoryMessage>
-      <ValidationRegExpMessage></ValidationRegExpMessage>
-      <LabelOnTop>true</LabelOnTop>
-    </GenericProperty>
-  </GenericProperties>
-  <Tabs>
-    <Tab>
-      <Key>881412d4-0666-4cc4-872f-bb41f3b8af09</Key>
-      <Caption>Settings</Caption>
-      <Alias>settings</Alias>
-      <Type>Group</Type>
-      <SortOrder>0</SortOrder>
-    </Tab>
-  </Tabs>
+	<Info>
+		<Name>Fieldset settings</Name>
+		<Icon>icon-equalizer color-black</Icon>
+		<Thumbnail>folder.png</Thumbnail>
+		<Description>Settings for the fieldset component. Use the fieldset component when you need to show a relationship between multiple form inputs.</Description>
+		<AllowAtRoot>False</AllowAtRoot>
+		<IsListView>False</IsListView>
+		<Variations>Nothing</Variations>
+		<IsElement>true</IsElement>
+		<HistoryCleanup>
+			<PreventCleanup>False</PreventCleanup>
+			<KeepAllVersionsNewerThanDays></KeepAllVersionsNewerThanDays>
+			<KeepLatestVersionPerDayForDays></KeepLatestVersionPerDayForDays>
+		</HistoryCleanup>
+		<Folder>GOV.UK</Folder>
+		<Compositions>
+			<Composition Key="bf69e8d6-97a6-4b50-bec7-c5993e120339">govukCssClasses</Composition>
+			<Composition Key="d508995b-a958-4273-8ab3-50735220a1b2">govukGrid</Composition>
+			<Composition Key="d7016563-0b6f-4c3a-b75d-b0bbe66419fd">govukGridColumnClasses</Composition>
+			<Composition Key="8ba759bf-131b-488e-b7e6-98ff91bb076b">govukLegendIsPageHeading</Composition>
+		</Compositions>
+		<DefaultTemplate></DefaultTemplate>
+		<AllowedTemplates />
+	</Info>
+	<Structure />
+	<GenericProperties>
+		<GenericProperty>
+			<Key>052c190a-53f3-400c-bbd7-41d69328aa35</Key>
+			<Name>Fieldset errors enabled</Name>
+			<Alias>fieldsetErrors</Alias>
+			<Definition>189de753-cac2-4307-868b-114832c68a07</Definition>
+			<Type>Umbraco.TrueFalse</Type>
+			<Mandatory>false</Mandatory>
+			<Validation></Validation>
+			<Description><![CDATA[When a child 'Error message' component is bound to a property on the view model which fails validation, that is a fieldset error.]]></Description>
+			<SortOrder>10</SortOrder>
+			<Tab Alias="settings">Settings</Tab>
+			<Variations>Nothing</Variations>
+			<MandatoryMessage></MandatoryMessage>
+			<ValidationRegExpMessage></ValidationRegExpMessage>
+			<LabelOnTop>true</LabelOnTop>
+		</GenericProperty>
+	</GenericProperties>
+	<Tabs>
+		<Tab>
+			<Key>881412d4-0666-4cc4-872f-bb41f3b8af09</Key>
+			<Caption>Settings</Caption>
+			<Alias>settings</Alias>
+			<Type>Group</Type>
+			<SortOrder>0</SortOrder>
+		</Tab>
+	</Tabs>
 </ContentType>
