+++ GovUk.Frontend.Umbraco/Views/Partials/GOVUK/GovUkSelect.cshtml
@@ -3,7 +3,6 @@
 @addTagHelper *, GovUk.Frontend.AspNetCore.Extensions
 @using GovUk.Frontend.Umbraco
 @using GovUk.Frontend.Umbraco.Models
-@using GovUk.Frontend.Umbraco.Services
 @using GovUk.Frontend.Umbraco.Typography
 @using Microsoft.AspNetCore.Mvc.ModelBinding
 @using Umbraco.Cms.Core.Models.Blocks
@@ -20,11 +19,10 @@
     var attemptedValue = modelStateEntry?.AttemptedValue?.ToUpperInvariant();
     var labelIsPageHeading = Model.Settings.Value<bool>("labelIsPageHeading");
     var hint = GovUkTypography.Apply(Model.Content.Value<string>("hint"), new TypographyOptions { RemoveWrappingParagraph = true });
-    var describedBy = GovUkFieldsetAttributeBuilder.BuildAriaDescribedByForFieldsetErrors(Model.Settings.Properties, ViewContext.ModelState);
 }
 <govuk-client-side-validation
     error-message-required="@(Model.Settings.Value<string>(PropertyAliases.ErrorMessageRequired))">
-    <govuk-select name="@modelPropertyName" class="@cssClass" described-by="@describedBy">
+    <govuk-select name="@modelPropertyName" class="@cssClass">
         <govuk-select-label is-page-heading="@labelIsPageHeading" class="@(labelIsPageHeading ? "govuk-label--l" : null)">@(Model.Content.Value<string>("label"))</govuk-select-label>
         @if (!string.IsNullOrWhiteSpace(hint))
         {
